core/future: Implement continuation with make_task()
This patch removes struct continuation in favour of using make_task().
To do this, we store the type-erased, raw task function in the
promise. This allows for some simplifications at the expense of
additional space, as we're now storing a noncopyable_function. Note,
however, that we could make the internal noncopyable_function storage
size configurable, if this becomes an actual problem.

Note that keeping the continuation task object out of the promise
object would have required additional bookkeeping, as the objects
would have to reference each other. Promise would reference the task,
so it could be scheduled upon make_ready(); the task would have to
reference the promise, so as to update it's own reference when the
task would be moved (in future patches, we'll hold naked tasks in the
scheduler's data structures instead of pointers). Because the promise
can also be moved, it would need to inform the task of this. This
would make leveraging make_task() in schedule() very awkward, instead
of leveraging make_task() in make_ready() as implemented by this
patch.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>