core/reactor: Use chunked_fifo for task list
In subsequent patches we will store naked tasks, so using a
circular_buffer, which requires contiguous memory, would tax the
memory allocator too much. So, switch to a chunked_fifo.

A circular_buffer would also require task to be copyable, which it
currently isn't due to us relying on noncopyable_function.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>