core/gate: Add retryable with_gate primitive
Expose an overload of with_gate that accepts a retry policy, which can
reschedule the specified function. The retry policy is aware of the
gate so that the retry can be interrupted in case the gate is closed.

As with with_gate, the fiber only leaves the gate once it ultimately
resolves.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>